requirements.txt
asgiref==3.8.1
Django==5.1
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
graphviz==0.20.3
joblib==1.4.2
numpy==2.0.1
pandas==2.2.2
pillow==10.4.0
pip==24.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.1
